add various intro states to mobile. fixes missing loading screens
